Refactoring RviInsts to contain RV32 and RV32 inst handlers
