+++ core/inst_handlers/CMakeLists.txt
@@ -1,6 +1,12 @@
 project(Atlas)
 
-add_subdirectory(rv64)
+add_subdirectory(i)
+add_subdirectory(m)
+add_subdirectory(a)
+add_subdirectory(f)
+add_subdirectory(d)
+add_subdirectory(zicsr)
+add_subdirectory(zifencei)
 
 add_library(atlasinsts)
 target_link_libraries(atlasinsts
