+++ core/inst_handlers/d/CMakeLists.txt
