+++ core/inst_handlers/f/RvfInsts.cpp
@@ -1,4 +1,4 @@
-#include "core/inst_handlers/rv64/f/RvfInsts.hpp"
+#include "core/inst_handlers/f/RvfInsts.hpp"
 #include "include/ActionTags.hpp"
 #include "core/ActionGroup.hpp"
