+++ core/inst_handlers/i/RviInsts.hpp
@@ -14,8 +14,10 @@ namespace atlas
       public:
         using base_type = RviInsts;
 
-        static void getInstComputeAddressHandlers(std::map<std::string, Action> & inst_handlers_);
-        static void getInstHandlers(std::map<std::string, Action> & inst_handlers_);
+        template<typename XLEN>
+        static void getInstComputeAddressHandlers(std::map<std::string, Action> &);
+        template<typename XLEN>
+        static void getInstHandlers(std::map<std::string, Action> &);
 
       private:
         ActionGroup* add_64_handler(atlas::AtlasState* state);
