--- core/inst_handlers/rv64/CMakeLists.txt
@@ -1,9 +1,0 @@
-project(Atlas)
-
-add_subdirectory(i)
-add_subdirectory(m)
-add_subdirectory(a)
-add_subdirectory(f)
-add_subdirectory(d)
-add_subdirectory(zicsr)
-add_subdirectory(zifencei)
